requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-Babel==2.0.0
Flask-Talisman==1.1.0
Flask-Session==0.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Werkzeug==2.3.4
SQLAlchemy==2.0.12
gunicorn==20.1.0